+++ sok-example/src/main/kotlin/petshop/Swagger2KotlinExampleApplication.kt
@@ -1,0 +1,13 @@
+package petshop
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+
+@SpringBootApplication
+open class Swagger2KotlinExampleApplication
+
+
+fun main(args: Array<String>) {
+    SpringApplication.run(Swagger2KotlinExampleApplication::class.java, *args)
+}