+++ changelog.md
@@ -11,6 +11,10 @@ All notable changes to this project documented in this file.
 - Response definitions
 - Parameters definitions
 - List of Tags without `vendorExtensions` map
+- Rename to `sok` as `swagger on kotlin`
+- Add `hub` and `example` projects
+- Implemented core functionality for `hub`
+- Add `petshop.yml` as an example for `sok-example`
 ### Changed
 ### Deprecated
 ### Removed